backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "ApnaParivar - Family Tree Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.120.4",
    "python-dotenv>=1.0.1",
    "supabase>=2.4.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
